requirements.txt
# Data Processing
pandas>=1.5.0
numpy>=1.24.0

# Database connectivity
psycopg2-binary>=2.9.0

# Data visualization
matplotlib>=3.6.0
seaborn>=0.12.0

# Jupyter notebook support
jupyter>=1.0.0
ipython>=8.0.0

# Additional data processing utilities
openpyxl>=3.0.0  # For Excel file support
xlrd>=2.0.0      # For reading Excel files

# Optional: For enhanced data analysis
scikit-learn>=1.2.0  # For any ML preprocessing
plotly>=5.0.0        # For interactive visualizations

# Development and testing
pytest>=7.0.0        # For testing
black>=22.0.0        # For code formatting
flake8>=5.0.0        # For linting 